refactor: move apply_font_settings to plot_utils, call once at orchestration level
- Move apply_font_settings from plotter.py to plot_utils.py (clean imports)
- Call once in plot_orchestration.py instead of in every plotter __init__
- Remove redundant calls from Plotter, LinePlots, QuantilePlots, ScoreCards, BarPlots
