+++ packages/evaluate/src/weathergen/evaluate/plotting/plotter.py
@@ -72,26 +72,6 @@ logging.getLogger("matplotlib.category").setLevel(logging.ERROR)
 _logger.debug(f"Taking cartopy paths from {work_dir}")
 
 
-def apply_font_settings(cfg: dict) -> None:
-    """Apply font settings from plotter config to matplotlib rcParams.
-
-    Parameters
-    ----------
-    cfg : dict
-        Plotter configuration dictionary.  Recognised keys:
-
-        - ``font_size``: base font size (default: matplotlib default)
-        - ``font_type``: font family, e.g. ``'serif'``, ``'sans-serif'``,
-          ``'monospace'`` (default: matplotlib default)
-    """
-    font_size = cfg.get("font_size")
-    font_type = cfg.get("font_type")
-    if font_size is not None:
-        mpl.rcParams["font.size"] = font_size
-    if font_type is not None:
-        mpl.rcParams["font.family"] = font_type
-
-
 @dataclass
 class DistStats:
     """Summary statistics for a 1-D distribution."""
@@ -153,7 +133,6 @@ class Plotter:
 
         _logger.debug(f"Taking cartopy paths from {work_dir}")
 
-        apply_font_settings(plotter_cfg)
         self.image_format = plotter_cfg.get("image_format")
         self.animation_format = plotter_cfg.get("animation_format")
         self.dpi_val = plotter_cfg.get("dpi_val")
